(fix) auth middleware for jobs

+++ src/jobs/jobs.service.ts
@@ -1,4 +1,10 @@
-import { HttpException, HttpStatus, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CreateJobDtoSchema } from './dto/create-job.dto';
 import { UpdateJobDtoSchema } from './dto/update-job.dto';
 import { PrismaService } from 'src/shared/prisma/prisma.service';
@@ -54,12 +60,28 @@ export class JobsService {
     });
   }
 
-  async remove(idJob: string): Promise<JobPost> {
-    const job = (await this.findOne(idJob)) as Prisma.JobPostWhereUniqueInput;
+  async remove(userId: string, idJob: string): Promise<JobPost> {
+    const job = await this.prisma.jobPost.findUnique({
+      where: { id: idJob },
+    });
+
+    if (!job) {
+      throw new NotFoundException(`Job with ID "${idJob}" not found`);
+    }
+
+    if (job.idAluno !== userId) {
+      throw new UnauthorizedException(
+        "You don't have permission to update this job",
+      );
+    }
     return this.prisma.jobPost.delete({ where: job });
   }
 
-  async updateJob(userId: string, jobId: string, updateJobDto: UpdateJobDtoSchema) {
+  async updateJob(
+    userId: string,
+    jobId: string,
+    updateJobDto: UpdateJobDtoSchema,
+  ) {
     const job = await this.prisma.jobPost.findUnique({
       where: { id: jobId },
     });
@@ -69,7 +91,9 @@ export class JobsService {
     }
 
     if (job.idAluno !== userId) {
-      throw new UnauthorizedException("You don't have permission to update this job");
+      throw new UnauthorizedException(
+        "You don't have permission to update this job",
+      );
     }
 
     return this.prisma.jobPost.update({
